@@ -1,51 +1,53 @@
 {
   "name": "referee-coach-ui",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "build-staging": "ng build --configuration staging",
     "watch": "ng build --watch --configuration development",
-    "lint": "ng lint"
+    "lint": "ng lint",
+    "release": "release-it"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "19.1.5",
-    "@angular/cdk": "19.1.4",
-    "@angular/common": "19.1.5",
-    "@angular/compiler": "19.1.5",
-    "@angular/core": "19.1.5",
-    "@angular/forms": "19.1.5",
-    "@angular/material": "19.1.4",
-    "@angular/material-luxon-adapter": "19.1.4",
-    "@angular/platform-browser": "19.1.5",
-    "@angular/platform-browser-dynamic": "19.1.5",
-    "@angular/router": "19.1.5",
-    "@angular/youtube-player": "19.1.4",
+    "@angular/animations": "20.3.2",
+    "@angular/cdk": "20.2.5",
+    "@angular/common": "20.3.2",
+    "@angular/compiler": "20.3.2",
+    "@angular/core": "20.3.2",
+    "@angular/forms": "20.3.2",
+    "@angular/material": "20.2.5",
+    "@angular/material-luxon-adapter": "20.2.5",
+    "@angular/platform-browser": "20.3.2",
+    "@angular/platform-browser-dynamic": "20.3.2",
+    "@angular/router": "20.3.2",
+    "@angular/youtube-player": "20.2.5",
     "file-saver": "2.0.5",
     "get-video-id": "4.1.7",
-    "luxon": "3.5.0",
-    "rxjs": "7.8.1",
+    "luxon": "3.7.2",
+    "rxjs": "7.8.2",
     "tslib": "2.8.1",
-    "zone.js": "0.15.0"
+    "zone.js": "0.15.1"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "19.1.0",
-    "@angular-eslint/eslint-plugin": "19.1.0",
-    "@angular-eslint/eslint-plugin-template": "19.1.0",
-    "@angular-eslint/schematics": "19.1.0",
-    "@angular-eslint/template-parser": "19.1.0",
-    "@angular/build": "19.1.6",
-    "@angular/cli": "19.1.6",
-    "@angular/compiler-cli": "19.1.5",
+    "@angular-eslint/builder": "20.3.0",
+    "@angular-eslint/eslint-plugin": "20.3.0",
+    "@angular-eslint/eslint-plugin-template": "20.3.0",
+    "@angular-eslint/schematics": "20.3.0",
+    "@angular-eslint/template-parser": "20.3.0",
+    "@angular/build": "20.3.3",
+    "@angular/cli": "20.3.3",
+    "@angular/compiler-cli": "20.3.2",
     "@types/file-saver": "2.0.7",
-    "@types/luxon": "3.4.2",
-    "@types/node": "22.13.1",
-    "@typescript-eslint/eslint-plugin": "8.24.0",
-    "@typescript-eslint/parser": "8.24.0",
-    "eslint": "9.20.1",
+    "@types/luxon": "3.7.1",
+    "@types/node": "24.5.2",
+    "@typescript-eslint/eslint-plugin": "8.44.1",
+    "@typescript-eslint/parser": "8.44.1",
+    "eslint": "9.36.0",
+    "release-it": "19.0.5",
     "schematics-scss-migrate": "2.3.17",
-    "typescript": "5.7.3"
+    "typescript": "5.9.2"
   }
 }
