@@ -1,5 +1,5 @@
-import {AfterViewInit, Component, ElementRef, HostListener, OnDestroy, OnInit, ViewChild} from '@angular/core';
-import {ActivatedRoute} from "@angular/router";
+import {AfterViewInit, Component, ElementRef, HostListener, inject, OnDestroy, OnInit, viewChild} from '@angular/core';
+import {ActivatedRoute, RouterLink} from "@angular/router";
 import {VideoReportService} from "../service/video-report.service";
 import {AuthenticationService} from "../service/authentication.service";
 import {MatDialog, MatDialogConfig} from "@angular/material/dialog";
@@ -11,18 +11,36 @@ import {Observable, of} from "rxjs";
 import {UnsavedRepliesDialogComponent} from "../discuss-video-report-unsaved-replies-dialog/unsaved-replies-dialog.component";
 import {DiscussVideoReportFinishDialogComponent} from "../discuss-video-report-finish-dialog/discuss-video-report-finish-dialog.component";
 import {HasUnsavedReplies} from "../has-unsaved-replies";
+import {MatToolbar} from '@angular/material/toolbar';
+import {DatePipe, NgClass} from '@angular/common';
+import {MatProgressSpinner} from '@angular/material/progress-spinner';
+import {MatButton, MatIconAnchor, MatIconButton} from '@angular/material/button';
+import {MatIcon} from '@angular/material/icon';
+import {MatTooltip} from '@angular/material/tooltip';
+import {MatProgressBar} from '@angular/material/progress-bar';
+import {MatCard, MatCardActions, MatCardContent, MatCardHeader, MatCardTitle} from '@angular/material/card';
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {CdkTextareaAutosize} from '@angular/cdk/text-field';
+import {MatInput} from '@angular/material/input';
+import {FormsModule} from '@angular/forms';
 
 @Component({
     selector: 'app-discuss-video-report',
     templateUrl: './discuss-video-report.component.html',
     styleUrls: ['./discuss-video-report.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatProgressSpinner, MatIconButton, MatIcon, MatTooltip, MatIconAnchor, RouterLink, MatProgressBar, MatCard, MatCardHeader, MatCardTitle, MatCardContent, YouTubePlayer, MatButton, NgClass, MatFormField, MatLabel, CdkTextareaAutosize, MatInput, FormsModule, MatCardActions, DatePipe]
 })
 export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDestroy, HasUnsavedReplies {
+    private route = inject(ActivatedRoute);
+    private videoReportService = inject(VideoReportService);
+    private authenticationService = inject(AuthenticationService);
+    dialog = inject(MatDialog);
+    snackBar = inject(MatSnackBar);
 
-    @ViewChild('youtubePlayer') youtube?: YouTubePlayer;
-    @ViewChild('widthMeasurement') widthMeasurement?: ElementRef<HTMLDivElement>;
-    @ViewChild('videoCommentsContainer') videoCommentsContainer?: ElementRef<HTMLDivElement>;
+
+    readonly youtube = viewChild<YouTubePlayer>('youtubePlayer');
+    readonly widthMeasurement = viewChild<ElementRef<HTMLDivElement>>('widthMeasurement');
+    readonly videoCommentsContainer = viewChild<ElementRef<HTMLDivElement>>('videoCommentsContainer');
 
     videoWidth?: number;
     videoHeight?: number;
@@ -33,13 +51,6 @@ export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDes
 
     replies: CommentReplyDTO[] = [];
     newComments: VideoCommentDTO[] = [];
-
-    constructor(private route: ActivatedRoute,
-                private videoReportService: VideoReportService,
-                private authenticationService: AuthenticationService,
-                public dialog: MatDialog,
-                public snackBar: MatSnackBar) {
-    }
 
     ngOnInit(): void {
         // This code loads the IFrame Player API code asynchronously, according to the instructions at
@@ -65,7 +76,7 @@ export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDes
 
     onResize = (): void => {
         // minus padding (16px each side) and margin (10px each)
-        const contentWidth = this.widthMeasurement!.nativeElement.clientWidth - 52;
+        const contentWidth = this.widthMeasurement()!.nativeElement.clientWidth - 52;
 
         this.videoWidth = Math.min(contentWidth, 720);
         this.videoHeight = this.videoWidth * 0.6;
@@ -103,8 +114,8 @@ export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDes
     }
 
     play(time: number): void {
-        this.youtube!.seekTo(time, true);
-        this.youtube!.playVideo();
+        this.youtube()!.seekTo(time, true);
+        this.youtube()!.playVideo();
     }
 
     reply(comment: VideoCommentDTO): void {
@@ -176,7 +187,7 @@ export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDes
     }
 
     addVideoComment() {
-        const timestamp = Math.round(this.youtube!.getCurrentTime());
+        const timestamp = Math.round(this.youtube()!.getCurrentTime());
 
         // check if there is already a comment in the same timestamp range +/-3 seconds
         if (this.dto!.videoComments.some(comment => timestamp >= comment.timestamp - 3 && timestamp <= comment.timestamp + 3)) {
@@ -195,8 +206,9 @@ export class DiscussVideoReportComponent implements OnInit, AfterViewInit, OnDes
             this.newComments.push(newComment);
 
             setTimeout(() => {
-                if (this.videoCommentsContainer) {
-                    this.videoCommentsContainer.nativeElement.scrollTop = this.videoCommentsContainer.nativeElement.scrollHeight;
+                const videoCommentsContainer = this.videoCommentsContainer();
+                if (videoCommentsContainer) {
+                    videoCommentsContainer.nativeElement.scrollTop = videoCommentsContainer.nativeElement.scrollHeight;
                 }
             }, 200);
         }
