@@ -1,17 +1,30 @@
-import {Component, OnInit} from '@angular/core';
-import {FormBuilder, Validators} from '@angular/forms';
+import {Component, inject, OnInit} from '@angular/core';
+import {FormBuilder, FormsModule, ReactiveFormsModule, Validators} from '@angular/forms';
 import {AuthenticationService} from "../service/authentication.service";
-import {ActivatedRoute, Router} from "@angular/router";
+import {ActivatedRoute, Router, RouterLink} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {FORGOT_PASSWORD_PATH} from "../app-routing.module";
+import {MatToolbar} from '@angular/material/toolbar';
+
+import {MatProgressBar} from '@angular/material/progress-bar';
+import {MatCard, MatCardContent, MatCardHeader, MatCardTitle} from '@angular/material/card';
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {MatInput} from '@angular/material/input';
+import {MatButton} from '@angular/material/button';
 
 @Component({
     selector: 'app-login',
     templateUrl: './login.component.html',
     styleUrls: ['./login.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatProgressBar, MatCard, MatCardHeader, MatCardTitle, MatCardContent, FormsModule, ReactiveFormsModule, MatFormField, MatLabel, MatInput, MatButton, RouterLink]
 })
 export class LoginComponent implements OnInit {
+    private formBuilder = inject(FormBuilder);
+    private router = inject(Router);
+    private route = inject(ActivatedRoute);
+    private authenticationService = inject(AuthenticationService);
+    private snackBar = inject(MatSnackBar);
+
 
     authenticating = false;
 
@@ -21,13 +34,6 @@ export class LoginComponent implements OnInit {
         email: ['', [Validators.required]],
         password: ['', [Validators.required]],
     });
-
-    constructor(private formBuilder: FormBuilder,
-                private router: Router,
-                private route: ActivatedRoute,
-                private authenticationService: AuthenticationService,
-                private snackBar: MatSnackBar) {
-    }
 
     ngOnInit(): void {
         const email = this.route.snapshot.paramMap.get('email');
