@@ -1,28 +1,52 @@
-import {AfterViewInit, Component, ElementRef, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, inject, OnDestroy, OnInit, viewChild} from '@angular/core';
 import {VideoReportService} from "../service/video-report.service";
 import {TagDTO, VideoCommentDetailDTO} from "../rest";
-import {MatTableDataSource} from "@angular/material/table";
+import {
+    MatCell,
+    MatCellDef,
+    MatColumnDef,
+    MatHeaderCell,
+    MatHeaderCellDef,
+    MatHeaderRow,
+    MatHeaderRowDef,
+    MatNoDataRow,
+    MatRow,
+    MatRowDef,
+    MatTable,
+    MatTableDataSource
+} from "@angular/material/table";
 import {YouTubePlayer} from "@angular/youtube-player";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {Observable, of, share} from "rxjs";
+import {MatToolbar} from '@angular/material/toolbar';
+import {MatButton, MatIconAnchor, MatIconButton} from '@angular/material/button';
+import {RouterLink} from '@angular/router';
+import {MatIcon} from '@angular/material/icon';
+import {MatCard, MatCardActions, MatCardContent} from '@angular/material/card';
+import {TagsSelectionComponent} from '../tags-selection/tags-selection.component';
+import {DatePipe} from '@angular/common';
+import {MatProgressBar} from '@angular/material/progress-bar';
 
 @Component({
     selector: 'app-report-search',
     templateUrl: './report-search.component.html',
     styleUrls: ['./report-search.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatIconAnchor, RouterLink, MatIcon, MatCard, MatCardContent, TagsSelectionComponent, YouTubePlayer, MatCardActions, MatButton, MatProgressBar, MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatIconButton, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow, MatNoDataRow, MatPaginator, DatePipe]
 })
 export class ReportSearchComponent implements OnInit, AfterViewInit, OnDestroy {
+    private readonly videoReportService = inject(VideoReportService);
+    private snackBar = inject(MatSnackBar);
+
 
     displayedColumns: string[] = ['date', 'gameNumber', 'competition', 'comment', 'tags', 'play'];
 
-    @ViewChild('youtubePlayer') youtube?: YouTubePlayer;
-    @ViewChild('widthMeasurement') widthMeasurement?: ElementRef<HTMLDivElement>;
+    readonly youtube = viewChild<YouTubePlayer>('youtubePlayer');
+    readonly widthMeasurement = viewChild<ElementRef<HTMLDivElement>>('widthMeasurement');
 
     selectedTags: TagDTO[] = [];
     results: MatTableDataSource<VideoCommentDetailDTO> = new MatTableDataSource<VideoCommentDetailDTO>([]);
-    @ViewChild(MatPaginator) paginator: MatPaginator | undefined;
+    readonly paginator = viewChild(MatPaginator);
 
     currentVideoId?: string;
     videoWidth?: number;
@@ -31,10 +55,6 @@ export class ReportSearchComponent implements OnInit, AfterViewInit, OnDestroy {
     searching = false;
 
     availableTags: Observable<TagDTO[]> = of([]);
-
-    constructor(private readonly videoReportService: VideoReportService,
-                private snackBar: MatSnackBar) {
-    }
 
     ngOnInit(): void {
         // This code loads the IFrame Player API code asynchronously, according to the instructions at
@@ -70,8 +90,9 @@ export class ReportSearchComponent implements OnInit, AfterViewInit, OnDestroy {
         }).subscribe({
             next: response => {
                 this.results = new MatTableDataSource<VideoCommentDetailDTO>(response.results);
-                if (this.paginator) {
-                    this.results.paginator = this.paginator
+                const paginator = this.paginator();
+                if (paginator) {
+                    this.results.paginator = paginator
                 }
             },
             error: () => {
@@ -92,9 +113,10 @@ export class ReportSearchComponent implements OnInit, AfterViewInit, OnDestroy {
         this.currentVideoId = element.youtubeId;
 
         const interval = setInterval(() => {
-            if (this.youtube!.getPlayerState() !== YT.PlayerState.UNSTARTED) {
-                this.youtube!.seekTo(element.timestamp, true);
-                this.youtube!.playVideo();
+            const youtube = this.youtube();
+            if (youtube!.getPlayerState() !== YT.PlayerState.UNSTARTED) {
+                youtube!.seekTo(element.timestamp, true);
+                youtube!.playVideo();
                 clearInterval(interval);
             }
         }, 500);
@@ -103,7 +125,7 @@ export class ReportSearchComponent implements OnInit, AfterViewInit, OnDestroy {
     onResize = (): void => {
         setTimeout(() => {
             // minus padding (16px each side) and margin (10px each)
-            const contentWidth = this.widthMeasurement!.nativeElement.clientWidth - 52;
+            const contentWidth = this.widthMeasurement()!.nativeElement.clientWidth - 52;
 
             this.videoWidth = Math.min(contentWidth, 720);
             this.videoHeight = this.videoWidth * 0.6;
