@@ -1,16 +1,29 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, inject, OnInit} from '@angular/core';
 import {AuthenticationService} from "../service/authentication.service";
-import {ActivatedRoute, Router} from "@angular/router";
+import {ActivatedRoute, Router, RouterLink} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {LOGIN_PATH} from "../app-routing.module";
+import {MatToolbar} from '@angular/material/toolbar';
+
+import {MatProgressBar} from '@angular/material/progress-bar';
+import {MatCard, MatCardContent} from '@angular/material/card';
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {MatInput} from '@angular/material/input';
+import {FormsModule} from '@angular/forms';
+import {MatButton} from '@angular/material/button';
 
 @Component({
     selector: 'app-reset-password',
     templateUrl: './reset-password.component.html',
     styleUrls: ['./reset-password.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatProgressBar, MatCard, MatCardContent, MatFormField, MatLabel, MatInput, FormsModule, MatButton, RouterLink]
 })
 export class ResetPasswordComponent implements OnInit {
+    private authenticationService = inject(AuthenticationService);
+    private router = inject(Router);
+    private route = inject(ActivatedRoute);
+    private snackBar = inject(MatSnackBar);
+
     loginUrl = '';
     email?: string | null;
     token?: string | null;
@@ -19,12 +32,6 @@ export class ResetPasswordComponent implements OnInit {
 
     processing = false;
     success = false;
-
-    constructor(private authenticationService: AuthenticationService,
-                private router: Router,
-                private route: ActivatedRoute,
-                private snackBar: MatSnackBar) {
-    }
 
     ngOnInit(): void {
         this.email = this.route.snapshot.paramMap.get('email');
