@@ -1,4 +1,4 @@
-import {Injectable} from '@angular/core';
+import {inject, Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {ChangePasswordRequestDTO, ForgotPasswordRequestDTO, LoginRequestDTO, LoginResponseDTO, ResetPasswordRequestDTO, UserRole} from "../rest";
 import {environment} from "../../environments/environment";
@@ -8,6 +8,8 @@ import {Observable} from "rxjs";
     providedIn: 'root'
 })
 export class AuthenticationService {
+    private readonly httpClient = inject(HttpClient);
+
 
     private readonly token = 'token';
     private readonly userId = 'user-id';
@@ -15,9 +17,6 @@ export class AuthenticationService {
     private readonly role = 'role';
 
     private baseUrl = environment.baseUrl;
-
-    constructor(private readonly httpClient: HttpClient) {
-    }
 
     login(email: string, password: string): Observable<LoginResponseDTO> {
         const request: LoginRequestDTO = {
