@@ -1,4 +1,4 @@
-import {Injectable} from '@angular/core';
+import {inject, Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {BasketplanGameDTO} from "../rest";
@@ -8,11 +8,10 @@ import {environment} from "../../environments/environment";
     providedIn: 'root'
 })
 export class BasketplanService {
+    private readonly httpClient = inject(HttpClient);
+
 
     private baseUrl = environment.baseUrl;
-
-    constructor(private readonly httpClient: HttpClient) {
-    }
 
     searchGame(gameNumber: string): Observable<BasketplanGameDTO> {
         return this.httpClient.get<BasketplanGameDTO>(`${this.baseUrl}/game/${gameNumber}`);
