@@ -1,4 +1,4 @@
-import {Injectable} from '@angular/core';
+import {inject, Injectable} from '@angular/core';
 import {DateTime} from "luxon";
 import {Observable} from "rxjs";
 import {OverviewDTO} from "../rest";
@@ -9,11 +9,10 @@ import {HttpClient} from "@angular/common/http";
     providedIn: 'root'
 })
 export class SearchService {
+    private readonly httpClient = inject(HttpClient);
+
 
     private baseUrl = environment.baseUrl;
-
-    constructor(private readonly httpClient: HttpClient) {
-    }
 
     find(from: DateTime, to: DateTime): Observable<OverviewDTO[]> {
         return this.httpClient.get<OverviewDTO[]>(`${this.baseUrl}/video-report?from=${from.toFormat('yyyy-MM-dd')}&to=${to.toFormat('yyyy-MM-dd')}`);
