@@ -1,14 +1,25 @@
-import {Component} from '@angular/core';
-import {FormBuilder, Validators} from "@angular/forms";
+import {Component, inject} from '@angular/core';
+import {FormBuilder, FormsModule, ReactiveFormsModule, Validators} from "@angular/forms";
 import {AuthenticationService} from "../service/authentication.service";
+import {MatToolbar} from '@angular/material/toolbar';
+import {MatButton, MatIconAnchor} from '@angular/material/button';
+import {RouterLink} from '@angular/router';
+import {MatIcon} from '@angular/material/icon';
+import {MatCard, MatCardContent, MatCardHeader, MatCardTitle} from '@angular/material/card';
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {MatInput} from '@angular/material/input';
+
 
 @Component({
     selector: 'app-settings',
     templateUrl: './settings.component.html',
     styleUrls: ['./settings.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatIconAnchor, RouterLink, MatIcon, MatCard, MatCardHeader, MatCardTitle, MatCardContent, FormsModule, ReactiveFormsModule, MatFormField, MatLabel, MatInput, MatButton]
 })
 export class SettingsComponent {
+    private formBuilder = inject(FormBuilder);
+    private authenticationService = inject(AuthenticationService);
+
 
     error = false;
     errorMessage = '';
@@ -19,10 +30,6 @@ export class SettingsComponent {
         newPassword1: [null, [Validators.required]],
         newPassword2: [null, [Validators.required]],
     });
-
-    constructor(private formBuilder: FormBuilder,
-                private authenticationService: AuthenticationService) {
-    }
 
     changePassword() {
         if (this.changePasswordForm.valid) {
