@@ -1,6 +1,10 @@
-import {Component, Inject, OnInit} from '@angular/core';
+import {Component, inject, OnInit} from '@angular/core';
 import {Reportee, UserDTO} from "../rest";
-import {MAT_DIALOG_DATA} from "@angular/material/dialog";
+import {MAT_DIALOG_DATA, MatDialogActions, MatDialogClose, MatDialogContent, MatDialogTitle} from "@angular/material/dialog";
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {MatOption, MatSelect} from '@angular/material/select';
+
+import {MatButton} from '@angular/material/button';
 
 interface ReporteeSelection {
     reportee: Reportee,
@@ -20,18 +24,17 @@ export interface VideoReportCopyDialogData {
     selector: 'app-video-report-copy-dialog',
     templateUrl: './video-report-copy-dialog.component.html',
     styleUrls: ['./video-report-copy-dialog.component.scss'],
-    standalone: false
+    imports: [MatDialogTitle, MatDialogContent, MatFormField, MatLabel, MatSelect, MatOption, MatDialogActions, MatButton, MatDialogClose]
 })
 export class VideoReportCopyDialogComponent implements OnInit {
+    data = inject<VideoReportCopyDialogData>(MAT_DIALOG_DATA);
+
 
     title = '';
     description = '';
 
     reportee?: Reportee
     reportees: ReporteeSelection[] = []
-
-    constructor(@Inject(MAT_DIALOG_DATA) public data: VideoReportCopyDialogData) {
-    }
 
     ngOnInit(): void {
         this.title = this.data.title;
