@@ -1,8 +1,8 @@
-import {AfterViewInit, Component, ElementRef, HostListener, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, HostListener, inject, OnDestroy, OnInit, viewChild} from '@angular/core';
 import {YouTubePlayer} from "@angular/youtube-player";
 import {VideoReportService} from "../service/video-report.service";
 import {OfficiatingMode, Reportee, TagDTO, VideoCommentDTO, VideoReportDTO} from "../rest";
-import {ActivatedRoute, Router} from "@angular/router";
+import {ActivatedRoute, Router, RouterLink} from "@angular/router";
 import {MatDialog} from "@angular/material/dialog";
 import {VideoReportFinishDialogComponent} from "../video-report-finish-dialog/video-report-finish-dialog.component";
 import {MatSnackBar} from "@angular/material/snack-bar";
@@ -10,18 +10,39 @@ import {Observable, of, share} from "rxjs";
 import {VideoReportUnsavedChangesDialogComponent} from "../video-report-unsaved-changes-dialog/video-report-unsaved-changes-dialog.component";
 import {VIEW_PATH} from "../app-routing.module";
 import {VideoReportCopyDialogComponent, VideoReportCopyDialogData} from "../video-report-copy-dialog/video-report-copy-dialog.component";
+import {MatToolbar} from '@angular/material/toolbar';
+import {DatePipe, DecimalPipe, NgClass} from '@angular/common';
+import {MatProgressSpinner} from '@angular/material/progress-spinner';
+import {MatButton, MatIconAnchor, MatIconButton} from '@angular/material/button';
+import {MatIcon} from '@angular/material/icon';
+import {MatTooltip} from '@angular/material/tooltip';
+import {MatProgressBar} from '@angular/material/progress-bar';
+import {MatCard, MatCardActions, MatCardContent, MatCardHeader, MatCardTitle} from '@angular/material/card';
+import {MatFormField, MatLabel} from '@angular/material/form-field';
+import {CdkTextareaAutosize} from '@angular/cdk/text-field';
+import {MatInput} from '@angular/material/input';
+import {FormsModule} from '@angular/forms';
+import {VideoReportRatingComponent} from '../video-report-rating/video-report-rating.component';
+import {MatCheckbox} from '@angular/material/checkbox';
+import {TagsSelectionComponent} from '../tags-selection/tags-selection.component';
 
 @Component({
     selector: 'app-video-report',
     templateUrl: './video-report.component.html',
     styleUrls: ['./video-report.component.scss'],
-    standalone: false
+    imports: [MatToolbar, MatProgressSpinner, MatIconButton, MatIcon, MatTooltip, MatIconAnchor, RouterLink, MatProgressBar, MatCard, MatCardHeader, MatCardTitle, MatCardContent, NgClass, MatFormField, MatLabel, CdkTextareaAutosize, MatInput, FormsModule, VideoReportRatingComponent, YouTubePlayer, MatButton, MatCheckbox, TagsSelectionComponent, MatCardActions, DecimalPipe, DatePipe]
 })
 export class VideoReportComponent implements OnInit, AfterViewInit, OnDestroy {
+    private readonly videoReportService = inject(VideoReportService);
+    private route = inject(ActivatedRoute);
+    private router = inject(Router);
+    dialog = inject(MatDialog);
+    snackBar = inject(MatSnackBar);
 
-    @ViewChild('youtubePlayer') youtube?: YouTubePlayer;
-    @ViewChild('widthMeasurement') widthMeasurement?: ElementRef<HTMLDivElement>;
-    @ViewChild('videoCommentsContainer') videoCommentsContainer?: ElementRef<HTMLDivElement>;
+
+    readonly youtube = viewChild<YouTubePlayer>('youtubePlayer');
+    readonly widthMeasurement = viewChild<ElementRef<HTMLDivElement>>('widthMeasurement');
+    readonly videoCommentsContainer = viewChild<ElementRef<HTMLDivElement>>('videoCommentsContainer');
 
     videoWidth?: number;
     videoHeight?: number;
@@ -32,13 +53,6 @@ export class VideoReportComponent implements OnInit, AfterViewInit, OnDestroy {
     saving = false;
 
     availableTags: Observable<TagDTO[]> = of([]);
-
-    constructor(private readonly videoReportService: VideoReportService,
-                private route: ActivatedRoute,
-                private router: Router,
-                public dialog: MatDialog,
-                public snackBar: MatSnackBar) {
-    }
 
     ngOnInit(): void {
         // This code loads the IFrame Player API code asynchronously, according to the instructions at
@@ -74,7 +88,7 @@ export class VideoReportComponent implements OnInit, AfterViewInit, OnDestroy {
 
     onResize = (): void => {
         // minus padding (16px each side) and margin (10px each)
-        const contentWidth = this.widthMeasurement!.nativeElement.clientWidth - 52;
+        const contentWidth = this.widthMeasurement()!.nativeElement.clientWidth - 52;
 
         this.videoWidth = Math.min(contentWidth, 720);
         this.videoHeight = this.videoWidth * 0.6;
@@ -92,8 +106,8 @@ export class VideoReportComponent implements OnInit, AfterViewInit, OnDestroy {
     }
 
     jumpTo(time: number): void {
-        this.youtube!.seekTo(time, true);
-        this.youtube!.playVideo();
+        this.youtube()!.seekTo(time, true);
+        this.youtube()!.playVideo();
     }
 
     save() {
@@ -182,14 +196,15 @@ export class VideoReportComponent implements OnInit, AfterViewInit, OnDestroy {
             id: undefined,
             comment: '',
             requiresReply: false,
-            timestamp: Math.round(this.youtube!.getCurrentTime()),
+            timestamp: Math.round(this.youtube()!.getCurrentTime()),
             replies: [],
             tags: []
         });
 
         setTimeout(() => {
-            if (this.videoCommentsContainer) {
-                this.videoCommentsContainer.nativeElement.scrollTop = this.videoCommentsContainer.nativeElement.scrollHeight;
+            const videoCommentsContainer = this.videoCommentsContainer();
+            if (videoCommentsContainer) {
+                videoCommentsContainer.nativeElement.scrollTop = videoCommentsContainer.nativeElement.scrollHeight;
             }
         }, 200);
     }
